gpu: keep AffineChunkTrim on-device during pulsified encoder translation
AffineChunkTrim had no CUDA/Metal lowering, so every occurrence forced a
device->host->device round trip. In the nemotron encoders (3.5 and EN)
it sits in the per-layer self-attention path, ~4 times per layer, so a
32-layer encoder was ping-ponging through the host on every pulse step
despite the transferred tensors being tiny.

Move AffineChunkTrim into tract-pulse-opl (alongside Delay/PulsePad/
PulsedRange) so tract-gpu can depend on it without pulling in the
pulsifier crate, and add GpuAffineChunkTrim, a device-resident
copy_nd-based translation wired into try_make_copy_based_op (shared by
CUDA and Metal). Tighten the EN model harness's enc_assert accordingly.
